pyproject.toml
[tool.poetry]
name = "junction"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
setuptools = "^75.4.0"
redis = "^5.2.0"
python-dotenv = "^1.0.1"
strawberry-graphql = {extras = ["fastapi"], version = "^0.248.1"}
httpx = "^0.27.2"
apscheduler = "^3.10.4"
dijkstar = "^2.6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
ruff = "^0.7.4"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
